please show the whole code
please show all the lines of the code. This is in total round about 400 lines.

please show the last hundred lines of code. I believe you have a line limit of three hundred and fifty-seven.

remark: copy the first 357 lines and add then the from the last 100 lines the both last functions drawButton and mousePressed is the whole code, which is current 400 lines.